+++ src/app/shared/shared.module.ts
@@ -2,17 +2,22 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HoverDirective } from './directives/hover/hover.directive';
 import { ColorClickDirective } from './directives/color-click/color-click.directive';
+import { OrderByPipe } from './pipes/order-by.pipe';
 
 
 
 @NgModule({
-  declarations: [HoverDirective, ColorClickDirective],
+  declarations: [HoverDirective, ColorClickDirective, OrderByPipe],
   imports: [
     CommonModule
   ],
   exports: [
+    // directives
     HoverDirective,
-    ColorClickDirective
+    ColorClickDirective,
+
+    // pipes
+    OrderByPipe
     ]
 })
 export class SharedModule { }